RD-2090 Rename map file name to match component name
